add authen and author for app con

+++ Controllers/ApplicationController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using SqlSugar.DistributedSystem.Snowflake;
 using System.Diagnostics;
+using System.Security.Authentication;
 using WebAPI.Entity;
 using WebAPI.Model;
 using WebAPI.Service;
@@ -38,6 +39,19 @@ namespace WebAPI.Controllers
         public ResponseModel Get([FromQuery] Application application,
                                  [FromQuery] SysUser user)
         {
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                var currentUser = SessionService.GetSessionInfo(HttpContext.Session);
+                if (currentUser.Permission == 1)
+                {
+                    application.StudentID = currentUser.SysUserID;
+                    user.Permission = 1;
+                }
+                else if (currentUser.Permission == 2)
+                {
+                    application.StaffID = currentUser.SysUserID;
+                }
+            }
             try
             {
                 return new SuccessResponseModel()
@@ -56,6 +70,10 @@ namespace WebAPI.Controllers
         [HttpPost]
         public ResponseModel Update([FromBody] Application application)
         {
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 ApplicationService.Save(application);
@@ -72,6 +90,10 @@ namespace WebAPI.Controllers
         [HttpPost]
         public ResponseModel Submit([FromBody] Application application)
         {
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 return new SuccessResponseModel()
@@ -90,6 +112,10 @@ namespace WebAPI.Controllers
         [HttpDelete]
         public ResponseModel Delete([FromBody] Application application)
         {
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 ApplicationService.Delete(application);
@@ -106,6 +132,11 @@ namespace WebAPI.Controllers
         [HttpPost]
         public ResponseModel ChangeState([FromBody] Application application)
         {
+            AuthenticationService.Authorization(HttpContext.Session, 2);
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 ApplicationService.ChangeState(application);
@@ -122,6 +153,11 @@ namespace WebAPI.Controllers
         [HttpPost]
         public ResponseModel Approve([FromBody] Application application)
         {
+            AuthenticationService.Authorization(HttpContext.Session, 2);
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 ApplicationService.Approve(application);
@@ -138,6 +174,11 @@ namespace WebAPI.Controllers
         [HttpPost]
         public ResponseModel Assign([FromBody] Application application)
         {
+            AuthenticationService.Authorization(HttpContext.Session, 2);
+            if (!SessionService.isOwner(HttpContext.Session, application))
+            {
+                return new FailureResponseModel("not your application");
+            }
             try
             {
                 ApplicationService.Assign(application);
